generalize COSMOSSource class to work with any catalog
